requirements.txt
python=3.10.11
biopython
beautifulsoup4
lxml
requests
pandas
nglview
aiohttp